core/requirements.txt
google-adk>=1.0.0
google-genai>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.35.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
httpx>=0.27.0
python-dotenv>=1.0.0
google-cloud-firestore>=2.19.0
google-cloud-storage>=2.18.0
firebase-admin>=6.6.0
playwright>=1.44.0
readability-lxml>=0.8.1
e2b-code-interpreter>=1.0.0
twilio>=9.0.0
livekit-api>=1.0.7,<2
# NOTE: livekit-agents is here ONLY so the LiveKit CLI detects this as an agent project.
# The actual agent uses Dockerfile.agent which installs requirements-agent.txt instead.
livekit-agents[google]==1.4.3
